Parallel Frustum Cull / Light Assignment / Environment Volume Assignment  (#353)
* thread updates

* hyperthread pause update for arm

* architecture detection updated

* exec macro update

* parallel exec

* parallel and single imp for comparison

* task steeling queue exp 1

* best config so far

* del pstl for adding it as submod

* poolstl as submodule

* pool stl include update

* update patsh to poolstl

* remove unused codes

* optmiz

* cull fix

* use ntt*

* tk profile timer - no console command

* command

* logger parallel print fix

* profile timer works fine

* reset avg time added

* reset simplified

* destroy scene fix

* format the aabb to make it toolkit like

* light cache invalidation fixes.

* mutex for tklog

* parallel query is unified, light assignment wip

* pass box by ref

* light assignment uses aabb

* remove unused path

* bug fix

* component invalidate fix

* assign light as it was

* unused codes removed

* remove more unused render code

* picking fixes

* stats uses 64 bit integer for measuring allocated ram.

* billboard replace refactor

* refactor casts and update old engine scenes

* thumbnail wrong render fix

* send presorted lights to create render job

* cache access fix

* dir light selection fix, convert caches to raw ptr

* volume cache update fix

* selection fix

+++ ToolKit/Skeleton.cpp
@@ -17,8 +17,6 @@
 #include "ToolKit.h"
 #include "Util.h"
 
-
-
 namespace ToolKit
 {
 
@@ -48,7 +46,7 @@ namespace ToolKit
     glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_NEAREST);
     glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_NEAREST);
 
-    AddVRAMUsageInBytes(ptr->m_width * ptr->m_height * 16);
+    AddVRAMUsageInBytes(ptr->m_width * ptr->m_height * (uint64) 16);
 
     ptr->m_initiated = true;
     ptr->m_loaded    = true;
